Switch to CB in VLM by default.
