+++ src/cpp/src/llm_pipeline.cpp
@@ -22,80 +22,6 @@ bool with_cpu_sve();
 
 namespace genai {
 
-namespace {
-
-const std::string PA_BACKEND = "PA";
-const std::string SDPA_BACKEND = "SDPA";
-
-inline bool is_paged_attention_available() {
-#ifdef OPENVINO_ARCH_X86_64
-    return true;
-#elif defined OPENVINO_ARCH_ARM64
-    return with_cpu_sve();
-#else
-    return false;
-#endif
-}
-
-SchedulerConfig get_latency_oriented_scheduler_config() {
-    SchedulerConfig default_config;
-    default_config.max_num_batched_tokens = std::numeric_limits<size_t>::max(); // don't limit total batch size
-    default_config.enable_prefix_caching = true; // for better TTFT in chat scenarios
-    return default_config;
-}
-
-bool explicitly_requires_paged_attention(const ov::AnyMap& properties) {
-    auto attention_backend_it = properties.find("ATTENTION_BACKEND");
-
-    if (properties.find(ov::genai::scheduler_config.name()) != properties.end() ||
-        (attention_backend_it != properties.end() && attention_backend_it->second.as<std::string>() == PA_BACKEND)) {
-        if (is_paged_attention_available()) {
-            return true;
-        } else {
-            OPENVINO_THROW("Continuous batching backend requires PagedAttention operation support, which is available on x86_64 or ARM64 with SVE platforms only");
-        }
-    }
-    if (properties.find(utils::DRAFT_MODEL_ARG_NAME) != properties.end()) {
-        if (is_paged_attention_available()) {
-            return true;
-        } else {
-            OPENVINO_THROW("Speculative decoding requires PagedAttention operation support, which is available on x86_64 or ARM64 with SVE platforms only");
-        }
-    }
-    if (properties.find(ov::genai::prompt_lookup.name()) != properties.end()) {
-        if (is_paged_attention_available()) {
-            return true;
-        } else {
-            OPENVINO_THROW("Prompt lookup decoding requires PagedAttention operation support, which is available on x86_64 or ARM64 with SVE platforms only");
-        }
-    }
-    return false;
-}
-
-std::pair<ov::AnyMap, std::string> extract_attention_backend(const ov::AnyMap& external_properties) {
-    std::string attention_backend = PA_BACKEND;
-    ov::AnyMap properties = external_properties;
-
-    auto it = properties.find("ATTENTION_BACKEND");
-    if (it != properties.end()) {
-        attention_backend = it->second.as<std::string>();
-        OPENVINO_ASSERT(attention_backend == PA_BACKEND || attention_backend == SDPA_BACKEND,
-            "Attention backend must be either '", PA_BACKEND, "' or '", SDPA_BACKEND, "', got '", attention_backend, "'");
-        properties.erase(it);
-    }
-
-    if (explicitly_requires_paged_attention(external_properties)) {
-        OPENVINO_ASSERT(attention_backend == PA_BACKEND,
-            "User properties are conflicting: some of them requires PagedAttention backend, while 'ATTENTION_BACKEND' is set to 'SDPA'");
-    }
-
-    return {properties, attention_backend};
-};
-
-
-} // namespace
-
-
 std::pair<std::string, Any> streamer(StreamerVariant func) {
     if (auto streamer_obj = std::get_if<std::shared_ptr<StreamerBase>>(&func)) {
         return {utils::STREAMER_ARG_NAME, Any::make<std::shared_ptr<StreamerBase>>(*streamer_obj)};
@@ -154,11 +80,11 @@ ov::genai::LLMPipeline::LLMPipeline(
     const std::string& device,
     const ov::AnyMap& user_properties) {
     auto start_time = std::chrono::steady_clock::now();
-    auto [properties, attention_backend] = extract_attention_backend(user_properties);
+    auto [properties, attention_backend] = utils::extract_attention_backend(user_properties);
 
     // If CB is invoked explicitly, create CB adapter as is and re-throw in case if internal issues
-    if (explicitly_requires_paged_attention(user_properties)) {
-        auto [device_properties, scheduler_config] = utils::extract_scheduler_config(properties, get_latency_oriented_scheduler_config());
+    if (utils::explicitly_requires_paged_attention(user_properties)) {
+        auto [device_properties, scheduler_config] = utils::extract_scheduler_config(properties, utils::get_latency_oriented_scheduler_config());
         m_pimpl = std::make_unique<ContinuousBatchingAdapter>(models_path, tokenizer, scheduler_config, device, device_properties);
     }
 
@@ -174,7 +100,7 @@ ov::genai::LLMPipeline::LLMPipeline(
             // we need use CB only for x86, as for other architectures like arm64 or risc-v we can create Paged Attention based model
             // but cannot perform its inference later
 #ifdef OPENVINO_ARCH_X86_64
-            m_pimpl = std::make_unique<ContinuousBatchingAdapter>(models_path, tokenizer, get_latency_oriented_scheduler_config(), device, properties);
+            m_pimpl = std::make_unique<ContinuousBatchingAdapter>(models_path, tokenizer, utils::get_latency_oriented_scheduler_config(), device, properties);
 #endif
         } catch (ov::Exception&) {
             // ignore exceptions from PA
@@ -194,11 +120,11 @@ ov::genai::LLMPipeline::LLMPipeline(
     const ov::AnyMap& user_properties) {
     auto start_time = std::chrono::steady_clock::now();
 
-    auto [properties, attention_backend] = extract_attention_backend(user_properties);
+    auto [properties, attention_backend] = utils::extract_attention_backend(user_properties);
 
     // If CB is invoked explicitly, create CB adapter as is and re-throw in case if internal issues
-    if (explicitly_requires_paged_attention(user_properties)) {
-        auto [device_properties, scheduler_config] = utils::extract_scheduler_config(properties, get_latency_oriented_scheduler_config());
+    if (utils::explicitly_requires_paged_attention(user_properties)) {
+        auto [device_properties, scheduler_config] = utils::extract_scheduler_config(properties, utils::get_latency_oriented_scheduler_config());
         m_pimpl = std::make_unique<ContinuousBatchingAdapter>(models_path, scheduler_config, device, device_properties);
     }
 
@@ -214,7 +140,7 @@ ov::genai::LLMPipeline::LLMPipeline(
             // we need use CB only for x86, as for other architectures like arm64 or risc-v we can create Paged Attention based model
             // but cannot perform its inference later
 #ifdef OPENVINO_ARCH_X86_64
-            m_pimpl = std::make_unique<ContinuousBatchingAdapter>(models_path, get_latency_oriented_scheduler_config(), device, properties);
+            m_pimpl = std::make_unique<ContinuousBatchingAdapter>(models_path, utils::get_latency_oriented_scheduler_config(), device, properties);
 #endif
         } catch (ov::Exception&) {
             // ignore exceptions from PA
@@ -237,11 +163,11 @@ ov::genai::LLMPipeline::LLMPipeline(
     const ov::genai::GenerationConfig& generation_config) {
     auto start_time = std::chrono::steady_clock::now();
 
-    auto [properties, attention_backend] = extract_attention_backend(user_properties);
+    auto [properties, attention_backend] = utils::extract_attention_backend(user_properties);
 
     // If CB is invoked explicitly, create CB adapter as is and re-throw in case if internal issues
-    if (explicitly_requires_paged_attention(user_properties)) {
-        auto [device_properties, scheduler_config] = utils::extract_scheduler_config(properties, get_latency_oriented_scheduler_config());
+    if (utils::explicitly_requires_paged_attention(user_properties)) {
+        auto [device_properties, scheduler_config] = utils::extract_scheduler_config(properties, utils::get_latency_oriented_scheduler_config());
         m_pimpl = std::make_unique<ContinuousBatchingAdapter>(model_str, weights_tensor,
                                                               tokenizer, scheduler_config, device, device_properties, generation_config);
     }
@@ -268,7 +194,7 @@ ov::genai::LLMPipeline::LLMPipeline(
             // but cannot perform its inference later
 #ifdef OPENVINO_ARCH_X86_64
             m_pimpl = std::make_unique<ContinuousBatchingAdapter>(model_str, weights_tensor, tokenizer,
-                                                                  get_latency_oriented_scheduler_config(), device, properties, generation_config);
+                                                                  utils::get_latency_oriented_scheduler_config(), device, properties, generation_config);
 #endif
         } catch (ov::Exception&) {
             // ignore exceptions from PA
