+++ src/cpp/src/visual_language/pipeline.cpp
@@ -324,19 +324,39 @@ public:
 VLMPipeline::VLMPipeline(
     const std::filesystem::path& models_dir,
     const std::string& device,
-    const ov::AnyMap& properties
+    const ov::AnyMap& user_properties
 ) {
     auto start_time = std::chrono::steady_clock::now();
+    auto [properties, attention_backend] = utils::extract_attention_backend(user_properties);
 
-    if (properties.find(scheduler_config.name()) != properties.end() ||
-        properties.find(utils::DRAFT_MODEL_ARG_NAME) != properties.end() ||
-        properties.find(prompt_lookup.name()) != properties.end()) {
-        auto [plugin_config, scheduler_config] = utils::extract_scheduler_config(properties);
-        m_pimpl = std::make_unique<VLMContinuousBatchingAdapter>(models_dir, scheduler_config, device, plugin_config);
+    // If CB is invoked explicitly, create CB adapter as is and re-throw in case if internal issues
+    if (utils::explicitly_requires_paged_attention(properties)) {
+        auto [plugin_properties, scheduler_config] = utils::extract_scheduler_config(properties, utils::get_latency_oriented_scheduler_config());
+        m_pimpl = std::make_unique<VLMContinuousBatchingAdapter>(models_dir, scheduler_config, device, plugin_properties);
     }
-    else {
+
+    if (m_pimpl == nullptr && device == "NPU") {
         m_pimpl = std::make_unique<VLMPipelineImpl>(models_dir, device, properties);
     }
+
+    // try to call CB adapter one more time, but with safe guard to silent exception
+    if (m_pimpl == nullptr && attention_backend == PA_BACKEND) {
+        try {
+            auto [plugin_properties, scheduler_config] = utils::extract_scheduler_config(properties, utils::get_latency_oriented_scheduler_config());
+            // we need use CB only for x86, as for other architectures like arm64 or risc-v we can create Paged Attention based model
+            // but cannot perform its inference later
+#ifdef OPENVINO_ARCH_X86_64
+            m_pimpl = std::make_unique<VLMContinuousBatchingAdapter>(models_dir, scheduler_config, device, plugin_properties);
+#endif
+        } catch (ov::Exception&) {
+            // ignore exceptions from PA
+        }
+    }
+
+    if (m_pimpl == nullptr) {
+        m_pimpl = std::make_unique<VLMPipelineImpl>(models_dir, device, properties);
+    }
+
     auto stop_time = std::chrono::steady_clock::now();
     m_pimpl->set_load_time(std::chrono::duration_cast<std::chrono::milliseconds>(stop_time - start_time).count());
 }
@@ -346,17 +366,37 @@ VLMPipeline::VLMPipeline(
     const Tokenizer& tokenizer,
     const std::filesystem::path& config_dir_path,
     const std::string& device,
-    const ov::AnyMap& properties,
+    const ov::AnyMap& user_properties,
     const GenerationConfig& generation_config
 ) {
     auto start_time = std::chrono::steady_clock::now();
-    if (properties.find(scheduler_config.name()) != properties.end() ||
-        properties.find(utils::DRAFT_MODEL_ARG_NAME) != properties.end() ||
-        properties.find(prompt_lookup.name()) != properties.end()) {
-        auto [plugin_config, scheduler_config] = utils::extract_scheduler_config(properties);
-        m_pimpl = std::make_unique<VLMContinuousBatchingAdapter>(models_map, tokenizer, config_dir_path, scheduler_config, device, plugin_config, generation_config);
+    auto [properties, attention_backend] = utils::extract_attention_backend(user_properties);
+
+    // If CB is invoked explicitly, create CB adapter as is and re-throw in case if internal issues
+    if (utils::explicitly_requires_paged_attention(properties)) {
+        auto [plugin_properties, scheduler_config] = utils::extract_scheduler_config(properties, utils::get_latency_oriented_scheduler_config());
+        m_pimpl = std::make_unique<VLMContinuousBatchingAdapter>(models_map, tokenizer, config_dir_path, scheduler_config, device, plugin_properties, generation_config);
     }
-    else {
+
+    if (m_pimpl == nullptr && device == "NPU") {
+        m_pimpl = std::make_unique<VLMPipelineImpl>(models_map, tokenizer, config_dir_path, device, properties, generation_config);
+    }
+
+    // try to call CB adapter one more time, but with safe guard to silent exception
+    if (m_pimpl == nullptr && attention_backend == PA_BACKEND) {
+        try {
+            auto [plugin_properties, scheduler_config] = utils::extract_scheduler_config(properties, utils::get_latency_oriented_scheduler_config());
+            // we need use CB only for x86, as for other architectures like arm64 or risc-v we can create Paged Attention based model
+            // but cannot perform its inference later
+#ifdef OPENVINO_ARCH_X86_64
+            std::make_unique<VLMContinuousBatchingAdapter>(models_map, tokenizer, config_dir_path, scheduler_config, device, plugin_properties, generation_config);
+#endif
+        } catch (ov::Exception&) {
+            // ignore exceptions from PA
+        }
+    }
+
+    if (m_pimpl == nullptr) {
         m_pimpl = std::make_unique<VLMPipelineImpl>(models_map, tokenizer, config_dir_path, device, properties, generation_config);
     }
     auto stop_time = std::chrono::steady_clock::now();
